Update CLI to call out to utils
As part of #105
