+++ util/build-migrations/index.js
@@ -1,84 +1,53 @@
 'use strict';
 
 const _ = require('lodash');
-const pgp = require('pg-promise');
+const fs = require('fs');
+const path = require('path');
+const loadResourceModels = require('../load-resource-models');
+const validateResourceModel = require('../validate-resource-model');
+const normalizeModel = require('../normalize-model');
+const depGraph = require('../resource-dependency-graph');
+const buildMigration = require('./util/build-migration');
 
-// I need to update this to at least use pgp's formatting library, and possibly
-// also a templating language. This is bad, I know.
+// Returns an array of migrations to apply. Throws an Error if a migration
+// cannot be created.
+module.exports = function(resourcesDir) {
+  // Load our resources. This will sort them based on their file name,
+  // which won't work when we run migrations. This is because relationships
+  // require that the migrations be run in the correct order.
+  const resources = loadResourceModels(resourcesDir);
 
-const idAttr = 'id SERIAL PRIMARY KEY';
-
-function generateAttrRow(value, attrName) {
-  let nullable = '';
-  if (!value.nullable) {
-    nullable = 'NOT NULL';
+  if (!resources.length) {
+    return [];
   }
 
-  let defaultValue = '';
-  if (value.default) {
-    // SQL Injection could occur here from `defaultValue`
-    defaultValue = pgp.as.format(`DEFAULT $[defaultValue^]`, {
-      defaultValue: value.default
-    });
+  // Attempt to catch errors early on by validating the model that the
+  // user has input.
+  const invalidResources = _.reject(resources, validateResourceModel);
+
+  if (invalidResources.length) {
+    throw new Error('There were invalid resources');
   }
 
-  // SQL Injection could occur here from `type`
-  const query = pgp.as.format(`$[attrName~] $[type^] ${nullable} ${defaultValue}`, {
-    type: value.type,
-    attrName
-  });
+  // If everything checks out, then we can normalize our resource
+  // to get it into a common format.
+  const normalizedResources = resources.map(normalizeModel);
 
-  return `  ${query}`;
-}
+  // Sort the resources based on their relationships.
+  const migrationOrder = depGraph(normalizedResources);
 
-function getTriggers(resource) {
-  const triggers = [];
-  if (resource.meta.updated_at) {
-    const query = pgp.as.format(
-      `CREATE TRIGGER updated_at BEFORE UPDATE ON $[resourceName~]
-      FOR EACH ROW EXECUTE PROCEDURE updated_at();`, {
-        resourceName: resource.name
-      });
+  // Retrieve the function migration, which is common to all DBs
+  const migrationsDir = path.join(__dirname, 'util', 'functions-migration.sql');
+  const fnMigration = fs.readFileSync(migrationsDir, {encoding: 'utf8'});
 
-    triggers.push(query);
-  }
+  // Create our up migrations. We assume that the resource has never
+  // existed. Eventually, we will need to first diff the resource
+  // against the previous version, then pass that diff into a method
+  // to get the migration!
+  const migrations = migrationOrder.map(resource => buildMigration(resource));
 
-  return triggers;
-}
+  // Ensure that the function migration is added
+  migrations.unshift(fnMigration);
 
-// At the moment this only works for many-to-one relations. The others will
-// require more complex updates to my migration building procedure.
-function generateRelationRow(relation, columnBaseName) {
-  let nullable = '';
-  if (!relation.nullable) {
-    nullable = ' NOT NULL';
-  }
-
-  const query = pgp.as.format(`$[columnName~] INTEGER references $[tableName~](id) ${nullable}`, {
-    columnName: `${columnBaseName}_id`,
-    tableName: relation.resource
-  });
-
-  return `  ${query}`;
-}
-
-module.exports = function(resource) {
-  const triggers = getTriggers(resource);
-
-  const idAttrColumn = [idAttr];
-  const attrs = _.map(resource.attributes, generateAttrRow);
-  const meta = _.map(resource.meta, generateAttrRow);
-  const relations = _.map(resource.relations, generateRelationRow);
-
-  const allColumns = idAttrColumn.concat(attrs, meta, relations);
-
-  const createTableQuery = pgp.as.format('CREATE TABLE $[resourceName~]', {
-    resourceName: resource.name
-  });
-
-  return `${createTableQuery} (
-  ${allColumns.join(',\n')}
-);
-
-${triggers.join('\n\n')}`;
+  return migrations;
 };
